unner: Reduce logspam
This change performs a small-scale refactor on unner. The primary
purpose is to obviate the kludge path in which Quod Libet commands are
forced to return a no-op ``/bin/true'' command which is then printed
to stdout. By changing the return type of `init_command()` from
`Command` to `Option<Command>`, we can return `None` from the Quod Libet
command runner.